Move modal component to folder
